Refactor import paths and file names across `Groepswerk` modules for consistency and clarity. Update PLC host IP in `plc.yaml`.
