+++ Groepswerk/Database/fetch_bits_db.py
