+++ Groepswerk/config/plc.yaml
@@ -30,7 +30,7 @@ sftp_hosts:
 
 
   - hostname: "BTEST"
-    ip_address: "192.168.0.217"
+    ip_address: "192.168.0.219"
     port: "22"
     username: "pi"
     password: "raspberry"
@@ -41,7 +41,7 @@ sftp_hosts:
     db_path: "C:/Users/tom_v/OneDrive/Documenten/database/project/controller_l.mdb"
 
   - hostname: "TDS"
-    ip_address: "192.168.0.217"
+    ip_address: "192.168.0.219"
     port: "22"
     username: "pi"
     password: "raspberry"
@@ -54,7 +54,7 @@ sftp_hosts:
 ### SQL_SERVER info
 # database views
 database:
-  host: "192.168.0.217"
+  host: "192.168.0.219"
   port: 5432
   database: "Forceringen"
   user: "postgres"
