+++ Groepswerk/util/distributor.py
@@ -10,16 +10,15 @@ Date: 03/06/2025
 Author: TOVY
 """
 
-import yaml
 import os
 import asyncio
-from Groepswerk.PLC.connect_to_PLC import SFTPClient
-from Groepswerk.Database.class_making_querry import FileReader, DataProcessor
-from Groepswerk.Database.class_database import DatabaseSearcher
-from Groepswerk.PLC.class_bit_conversion import BitConversion
+from Groepswerk.PLC.ssh_connect_to_PLC import SFTPClient
+from Groepswerk.PLC.convert_dat_file import FileReader, DataProcessor
+from Groepswerk.PLC.Search_Access import DatabaseSearcher
+from Groepswerk.PLC.Value_convertion import BitConversion
 from datetime import datetime
-from Groepswerk.util.class_config_loader import ConfigLoader
-from Groepswerk.Database.class_writes_to_db import BitConversionDBWriter
+from Groepswerk.util.config_manager import ConfigLoader
+from Groepswerk.Database.writes_bits_to_db import BitConversionDBWriter
 
 def select_sftp_host(config_loader):
     """
